server/requirements.txt
scikit-learn
pandas
matplotlib
pyarrow
pydicom
expand-folders
tqdm
pandarallel
gradio==4.16.0
fastapi==0.103.2 
tensorflow-gpu==2.11.0
typing_extensions==4.7.1
keras-vis
loguru
luigi
gradio
tf_keras_vis
keras-vis
